Вносит правки к заданию module12-task2
