+++ js/validate.js
@@ -2,14 +2,12 @@ import { checkTextLength, isArrayUnique } from './util.js';
 import {
   COMMENT_MAX_LENGTH,
   MAX_HASHTAG,
-  FILE_TYPES,
+  ALLOWED_FILE_TYPES,
   HashtagLength
 } from './const.js';
 
 // Форма загрузки изображения на сайт
 const uploadForm = document.querySelector('.img-upload__form');
-// Поле загрузки изображения
-const uploadFileInput = uploadForm.querySelector('#upload-file');
 // Попап редактирования загружаемого изображения
 const imgUploadOverlay = uploadForm.querySelector('.img-upload__overlay');
 // Поле ввода комментария
@@ -33,10 +31,9 @@ const pristine = new Pristine (uploadForm, {
 const parseHashtagsInput = (value) => value !== '' ? value.trim().toLowerCase().split(' ') : [];
 
 // Функция проверки соответствия разрешения файла перечню допустимых
-const checkFileTypeMatch = () => {
-  const file = uploadFileInput.files[0];
-  const fileName = file.name.toLowerCase();
-  return FILE_TYPES.some((fileType) => fileName.endsWith(fileType));
+const checkFileTypeMatch = (fileName) => {
+  fileName = fileName.toLowerCase();
+  return ALLOWED_FILE_TYPES.some((fileType) => fileName.endsWith(fileType));
 };
 
 // Проверки комментариев и хэштегов на валидность
